Extract redemption execution into src/redemptions/execution.py
